// src/entities/User/model/consts/userConsts.ts
export enum UserRole {
  SUPER_ADMIN = 'Super-Admin',
  ADMIN = 'Admin',
}

export enum UserPermission {
  USERS_CREATE = 'users.create',
  USERS_READ = 'users.read',
  USERS_UPDATE = 'users.update',
  USERS_DELETE = 'users.delete',
  CATEGORIES_CREATE = 'categories.create',
  CATEGORIES_READ = 'categories.read',
  CATEGORIES_UPDATE = 'categories.update',
  CATEGORIES_DELETE = 'categories.delete',
  ITEMS_CREATE = 'items.create',
  ITEMS_READ = 'items.read',
  ITEMS_UPDATE = 'items.update',
  ITEMS_DELETE = 'items.delete',
  TYPES_CREATE = 'types.create',
  TYPES_READ = 'types.read',
  TYPES_UPDATE = 'types.update',
  TYPES_DELETE = 'types.delete',
  GROUPS_CREATE = 'groups.create',
  GROUPS_READ = 'groups.read',
  GROUPS_UPDATE = 'groups.update',
  GROUPS_DELETE = 'groups.delete',
  ROOM_TYPES_CREATE = 'room_types.create',
  ROOM_TYPES_READ = 'room_types.read',
  ROOM_TYPES_UPDATE = 'room_types.update',
  ROOM_TYPES_DELETE = 'room_types.delete',
  ROOMS_CREATE = 'rooms.create',
  ROOMS_READ = 'rooms.read',
  ROOMS_UPDATE = 'rooms.update',
  ROOMS_DELETE = 'rooms.delete',
  CONFIRMATIONS_CREATE = 'confirmations.create',
  CONFIRMATIONS_READ = 'confirmations.read',
  CONFIRMATIONS_UPDATE = 'confirmations.update',
  CONFIRMATIONS_DELETE = 'confirmations.delete',
  CUSTOMERS_CREATE = 'customers.create',
  CUSTOMERS_READ = 'customers.read',
  CUSTOMERS_UPDATE = 'customers.update',
  CUSTOMERS_DELETE = 'customers.delete',
  SUPPLIERS_CREATE = 'suppliers.create',
  SUPPLIERS_READ = 'suppliers.read',
  SUPPLIERS_UPDATE = 'suppliers.update',
  SUPPLIERS_DELETE = 'suppliers.delete',
  CHECKINS_CREATE = 'checkins.create',
  CHECKINS_READ = 'checkins.read',
  CHECKINS_UPDATE = 'checkins.update',
  CHECKINS_DELETE = 'checkins.delete',
  CHECKOUTS_CREATE = 'checkouts.create',
  CHECKOUTS_READ = 'checkouts.read',
  CHECKOUTS_UPDATE = 'checkouts.update',
  CHECKOUTS_DELETE = 'checkouts.delete',
  TRANSFERS_CREATE = 'transfers.create',
  TRANSFERS_READ = 'transfers.read',
  TRANSFERS_UPDATE = 'transfers.update',
  TRANSFERS_DELETE = 'transfers.delete',
  PERMISSIONS_READ = 'permissions.read',
  ACTIVITIES_READ = 'activities.read',
  ROLES_CREATE = 'roles.create',
  ROLES_READ = 'roles.read',
  ROLES_UPDATE = 'roles.update',
  ROLES_DELETE = 'roles.delete',
  RESPONSIBILITIES_CREATE = 'responsibilities.create',
  RESPONSIBILITIES_READ = 'responsibilities.read',
  RESPONSIBILITIES_UPDATE = 'responsibilities.update',
  RESPONSIBILITIES_DELETE = 'responsibilities.delete',
}